+++ apps/search/forms.py
@@ -1,11 +1,16 @@
-from haystack.forms import ModelSearchForm
+from django import forms
+
+from apps.search.models import SearchDocument
 
 
-class HighlightedSearchForm(ModelSearchForm):
-    def search(self):
-        sqs = super().search()
-        kwargs = {
-            "hl.simple.pre": '<span class="highlighted">',
-            "hl.simple.post": "</span>",
-        }
-        return sqs.highlight(**kwargs)
+class SearchForm(forms.Form):
+    models = forms.MultipleChoiceField(
+        choices=[
+            (
+                f"{model._meta.app_label}.{model._meta.model_name}",
+                model._meta.verbose_name_plural,
+            )
+            for model in SearchDocument.indexed_models
+        ],
+        widget=forms.CheckboxSelectMultiple,
+    )