+++ apps/search/utils.py
@@ -1,38 +1,6 @@
-from haystack.exceptions import NotHandled
-from haystack.constants import DEFAULT_ALIAS
-from haystack import connections
-
 from django.apps import apps
 
-
-def update_document(object_id):
-    model, index, instance = get_model_instance_index(object_id)
-    if not instance or not index:
-        return
-    index._get_backend(DEFAULT_ALIAS).update(index, [instance])
-
-
-def remove_document(object_id):
-    model, index, instance = get_model_instance_index(object_id)
-    if not index:
-        return
-    index.remove_object(object_id, using=DEFAULT_ALIAS)
-
-
-def get_model_instance_index(object_id):
-    parts = object_id.split(".")
-    pk = parts[-1]
-    model_name = parts[-2]
-    app_name = ".".join(parts[:-2])
-    model = apps.get_model(app_name, model_name)
-    try:
-        index = connections[DEFAULT_ALIAS].get_unified_index().get_index(model)
-    except NotHandled:
-        index = None
-    try:
-        return model, index, model.objects.get(pk=pk)
-    except model.DoesNotExist:
-        return model, index, None
+from apps.search.models import SearchDocument
 
 
 def get_model(model_path):
@@ -40,3 +8,27 @@ def get_model(model_path):
     model_name = parts[-1]
     app_name = ".".join(parts[:-1])
     return apps.get_model(app_name, model_name)
+
+
+def post_save_handler(sender, instance, **kwargs):
+    SearchDocument.reindex(instance)
+    for rel in getattr(sender, "RELATED_MODELS", []):
+        model = get_model(rel[0])
+        objects = model.objects.filter(**{f"{rel[1]}": instance.pk})
+        for obj in objects:
+            SearchDocument.reindex(obj)
+
+
+def pre_delete_handler(sender, instance, **kwargs):
+    SearchDocument.remove(instance)
+    for rel in getattr(sender, "RELATED_MODELS", []):
+        model = get_model(rel[0])
+        objects = model.objects.filter(**{f"{rel[1]}": instance.pk})
+        for obj in objects:
+            SearchDocument.reindex(obj)
+
+
+def update_parent_handler(sender, instance, **kwargs):
+    parent = getattr(sender, "PARENT_FIELD", None)
+    if parent:
+        SearchDocument.reindex(getattr(instance, parent))